v0.3.0 - Object Class and Lanterns
Object class is now superclass of Entity, and the new Lantern!
Lantern does absolutely nothing but glow. It will have purpose.
Also fixed a blending issue with the lights which caused a gradient to override another which is baaad... now they use additive blending when drawing to the texture so now they merge